recovery: make recovery independant of ROM's
-The fact that CM puts kernel modules on the system partition and
Samsung puts them in the initramfs causes recovery to be dependant
on matching kernel modules in the ROM.

-To solve this we move the modules that recovery needs (param.ko
& j4fs.ko) to the initramfs of recovery.

-Now there will be no need to flash another recovery to flash
another ROM.

NOTE: To accomplish this goal I implemented a custom bootimage mk.
The kernel is still built the exact same way. It was necessary
to include it in shbootimg.mk.

The way the recovery is built has  changed in the following ways:
Before: recovery and ramdisk were built at the same time and
when those were done the kernel was build and finally the
recovery.img.

Now: kernel and recovery are built at the same time and when those
were done param.ko and j4fs.ko are copied to initramfs, the ramdisk
is built and finally the recovery.img.